requirements.txt
streamlit==1.29.0
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.22
pypdf==3.17.4
python-dotenv==1.0.0
huggingface-hub==0.20.2
sentence-transformers==2.2.2